Added a wrapper script that takes care of start and stop of the ses-server
